# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PMG Group OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentBody"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Contact:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        isDecisionMaker:
          type: boolean
        authorityLevel:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - isDecisionMaker
        - status
        - createdAt
        - updatedAt
    CreateContactBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        title:
          type: string
        role:
          type: string
        companyId:
          type: integer
        isDecisionMaker:
          type: boolean
        authorityLevel:
          type: string
        linkedinUrl:
          type: string
        status:
          type: string
        notes:
          type: string
      required:
        - firstName
        - lastName
    UpdateContactBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        title:
          type: string
        role:
          type: string
        companyId:
          type: integer
        isDecisionMaker:
          type: boolean
        authorityLevel:
          type: string
        linkedinUrl:
          type: string
        status:
          type: string
        notes:
          type: string
    Lead:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        contactId:
          type: ["integer", "null"]
        contactName:
          type: ["string", "null"]
        source:
          type: string
        status:
          type: string
        priority:
          type: string
        fitScore:
          type: ["integer", "null"]
        confidenceScore:
          type: ["integer", "null"]
        painPoints:
          type: ["string", "null"]
        bestAngle:
          type: ["string", "null"]
        nextAction:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        qualifiedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - source
        - status
        - priority
        - createdAt
        - updatedAt
    CreateLeadBody:
      type: object
      properties:
        companyId:
          type: integer
        contactId:
          type: integer
        source:
          type: string
        status:
          type: string
        priority:
          type: string
        fitScore:
          type: integer
        confidenceScore:
          type: integer
        painPoints:
          type: string
        bestAngle:
          type: string
        nextAction:
          type: string
        assignedTo:
          type: string
        notes:
          type: string
      required:
        - source
    UpdateLeadBody:
      type: object
      properties:
        companyId:
          type: integer
        contactId:
          type: integer
        source:
          type: string
        status:
          type: string
        priority:
          type: string
        fitScore:
          type: integer
        confidenceScore:
          type: integer
        painPoints:
          type: string
        bestAngle:
          type: string
        nextAction:
          type: string
        assignedTo:
          type: string
        notes:
          type: string
...
    Document:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        type:
          type: string
        content:
          type: ["string", "null"]
        status:
          type: string
        version:
          type: integer
        createdBy:
          type: ["string", "null"]
        approvedBy:
          type: ["string", "null"]
        entityType:
          type: ["string", "null"]
        entityId:
          type: ["integer", "null"]
        tags:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - category
        - type
        - status
        - version
        - createdAt
        - updatedAt
    CreateDocumentBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        type:
          type: string
        content:
          type: string
        status:
          type: string
        createdBy:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
        tags:
          type: string
        notes:
          type: string
      required:
        - title
        - category
        - type
    UpdateDocumentBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        type:
          type: string
        content:
          type: string
        status:
          type: string
        approvedBy:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
        tags:
          type: string
        notes:
          type: string
...