# lib/api-zod/src/generated/types/createContactBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PMG Group OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContactBody {
  firstName: string;
  lastName: string;
  email?: string;
  phone?: string;
  title?: string;
  role?: string;
  companyId?: number;
  isDecisionMaker?: boolean;
  authorityLevel?: string;
  linkedinUrl?: string;
  status?: string;
  notes?: string;
}