# lib/api-zod/src/generated/types/updateLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PMG Group OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateLeadBody {
  companyId?: number;
  contactId?: number;
  source?: string;
  status?: string;
  priority?: string;
  fitScore?: number;
  confidenceScore?: number;
  painPoints?: string;
  bestAngle?: string;
  nextAction?: string;
  assignedTo?: string;
  notes?: string;
}